Moved ChiouYoungs2014ACME2019 in the right module
